Created Add Visit function, Simplified other feats
AddVisitsButton in its infancy. Doesn't log time or track duration or even display the student's name,
    it just displays the ID of the student in a TableView and logs it in the histories DB with
    12:00 AM as a placeholder startTime
Next step is to make it impossible for a new visit to start without the last one for that student ending.
    (In SQL terms, a given visit ID can only have one record on the table with NULL in the endTime field)

To implement this the way I wanted it to, I had to rollback some of the progress I made with searching up
    students through a lineEdit widget. Simply put, I couldn't find a way to select/type a student into
    the lineEdit and then recall its ID from the table. This is a lot simpler with a non-editable
    completer-less ComboBox. Since a student's ID in the comboBox has the same row index as in the database
    itself